+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/GetBoundingBoxes.java
@@ -8,7 +8,6 @@ import com.azure.ai.formrecognizer.models.FormTable;
 import com.azure.ai.formrecognizer.models.FormWord;
 import com.azure.ai.formrecognizer.models.OperationResult;
 import com.azure.ai.formrecognizer.models.RecognizedForm;
-import com.azure.ai.formrecognizer.models.TextContentType;
 import com.azure.core.credential.AzureKeyCredential;
 import com.azure.core.util.polling.SyncPoller;
 
@@ -71,8 +70,8 @@ public class GetBoundingBoxes {
                         // call to beginRecognizeCustomFormsFromUrl
                         // It is also a list of FormWords and FormLines, but in this example, we only deal with
                         // FormWords
-                        formTableCell.getElements().forEach(formContent -> {
-                            if (formContent.getTextContentType().equals(TextContentType.WORD)) {
+                        formTableCell.getTextContent().forEach(formContent -> {
+                            if (formContent instanceof FormWord) {
                                 FormWord formWordElement = (FormWord) (formContent);
                                 StringBuilder boundingBoxStr = new StringBuilder();
                                 if (formWordElement.getBoundingBox() != null) {