+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -1,7 +1,9 @@
 # Release History
 
 ## 1.0.0-beta.4 (Unreleased)
-
+### Breaking Changes
+- Rename `elements` property on model `FormTableCell` to `textContent`
+- Remove `TextContentType` model and use `instanceOf` to detect the FormContent type
 
 ## 1.0.0-beta.3 (2020-06-10)
 ### New Features